Android Perforce UnauthorizedAccessException Fix
I've updated the android editor script to only overwrite the AndroidManifest.xml file it if it needs to be overwritten.  I also make sure to checkout the file if the user is using one of Unity's built in Source Control Providers.  Which is very much needed if you’re using Perforce.