Set versions to match Unity
Not specifying a version number sets it to a low version instead. minSDKVersion was set to 1 and added the READ_PHONE_STATE permission. Lower build-tools versions than the one the Unity Editor comes with were being required when building for Android. compileSDKVersion is required to be defined inorder to build and has been changed to match the version specified in Unity